pyproject.toml
[project]
name = "somnia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cerebras-cloud-sdk>=1.28.0",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "pydantic>=2.11.2",
    "pydantic-ai>=0.0.52",
    "sentence-transformers>=4.0.2",
    "supabase>=2.15.0",
    "torch>=2.6.0",
    "vecs>=0.4.5",
]
